Split Anthropic entity
